third_party/blink/web_tests/http/tests/devtools/profiler/heap-snapshot-weak-dominator-expected.txt

Tests that weak references are ignored when dominators are calculated and that weak references won't affect object's retained size.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testWeakReferencesDoNotAffectRetainedSize

Profiler was disabled.
